third_party/WebKit/LayoutTests/bluetooth/descriptor/writeValue/write-updates-value.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('DisconnectingHealthThermometerAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['health_thermometer']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('health_thermometer'))
    .then(service => service.getCharacteristic('measurement_interval'))
    .then(characteristic => characteristic.getDescriptor(user_description.name))
    .then(descriptor => {
      assert_equals(descriptor.value, null);
      let textEncoder = new TextEncoder();
      let newValue = textEncoder.encode('foo');
      return descriptor.writeValue(newValue).then(() => {
        assert_array_equals(descriptor.value.buffer, newValue.buffer);
      })
    })
}, 'Succesful read should update descriptor\'s value.');
</script>
